adminapp/requirements.txt
Django==1.9.7
amqp==2.1.4
argparse==1.2.1
billiard==3.5.0.2
celery==4.0.2
defusedxml==0.4.1
distribute==0.6.24
django-tastypie==0.13.3
kombu==4.0.2
python-dateutil==2.6.0
python-mimeparse==1.6.0
pytz==2017.2
requests==2.13.0
six==1.10.0
vine==1.1.3
wsgiref==0.1.2
